applications/earnings_call_interpreter/requirements.txt
streamlit
pandas
numpy
requests
pydantic
python-dotenv
tqdm
beautifulsoup4
lxml
PyPDF2
sentence-transformers
chromadb
networkx
neo4j
openai
tabulate
google-genai
wrds
ollama